Update patch set 1
Patch Set 1:

> the spec file is outdated.
 > 
 > Removing coverage is directly connected with tests disabled. Once
 > tests are enabled, you should be requiring coverage.

The tests *should* be run and I agree that is a problem that they are not being run here. The coverage report is not generated in RDO - please read 

 > the spec file is outdated.
 > 
 > Removing coverage is directly connected with tests disabled. Once
 > tests are enabled, you should be requiring coverage.

The tests should be run and I completely agree this is a problem and needs fixing. However RDO does not generate a coverage report, hence the reason for the following project:

https://github.com/redhat-openstack/easyfix/issues/11

Could I ask you to comment directly on the github issue if you disagree so it gets wider input as what you are saying essentially invalidates updating ~30 spec files. Thanks!

Patch-set: 1
Label: Code-Review=0