Update obsoleted version of python-django-openstack-auth
There has been a 3.6.0 release of python-django-openstack-auth in
Pike [1]. For the upgrade to work correctly on Queens and later,
we need to update the obsoleted version.

Setting it to < 4.0.0-1 to avoid future issues if there are further
version bumps in Pike.

[1] - https://review.rdoproject.org/r/13114

Change-Id: I2d5662ad014588cddb2a3e9c269f32999d8444b6
(cherry picked from commit 08933707e4417d66b7c2e695cf6bc4363e5d056b)