Fix spec processing by automated tools
The lua macro in charge of setting the required version for
openstack-dashboard-theme when rhosp is set works fine as long as
the %version macro is set to a number. In some cases, it might
still be 'XXX' when the spec file is being processed, which causes
issue for automation tools.

This patch adds a fallback for those cases, allowing the spec to
be properly parsed by rhe rpm python bindings.

Also, the systemd_postun macro in %post was causing issues when
parsed, since it now requires some arguments.

Change-Id: I0dc71a43fc179644aa28e35d7b0571688cf68cd2