Drop requirements.txt as deps are handled manually
We are managing dependencies manually in spec so
we don't need requirements.txt as they can cause
conflicts.
For EL9 builds due to this it's looking for some
specific versions which are not available and deps
which are not runtime dep like "enmerkar".

Change-Id: I67b2dd9711d92ca602bf992d27e6e8874003c72d
(cherry picked from commit 5ef305f698b5c5c1cb525da964d764d369c2c7f5)