Set python3-django requirement to < 3.0.0
While we tend to avoid setting upper bounds for requirements, the
Django requirement for Horizon is a bit special, since version 3.x
has been packaged in Fedora for some time [1], and it could end up
being included in some of the distros we rely on.

Setting this upper bound will allow us to identify this situation if
it happens, so we can make sure we are using the right requirement
version.

[1] - https://koji.fedoraproject.org/koji/packageinfo?packageID=13608

Change-Id: I6685470a65fca3a60a0d00987bc35ae41102d079